Abstract usage of `FiltersModal` into a component
